refactor(soul): accordion primitive
