+++ apps/web/vibes/soul/sections/products-list-section/filters-panel.tsx
@@ -5,7 +5,7 @@ import { Suspense, use } from 'react'
 import clsx from 'clsx'
 import { parseAsArrayOf, parseAsInteger, parseAsString, useQueryStates } from 'nuqs'
 
-import { Accordion, Accordions } from '@/vibes/soul/primitives/accordions'
+import * as Accordion from '@/vibes/soul/primitives/accordion'
 import { Button } from '@/vibes/soul/primitives/button'
 
 import { FilterRange } from './filter-range'
@@ -80,22 +80,22 @@ export function FiltersPanelInner({ className, filters }: Props) {
 
   return (
     <div className={clsx('w-full space-y-5', className)}>
-      <Accordions type="multiple" defaultValue={resolved.map((_, i) => i.toString())}>
+      <Accordion.Root type="multiple" defaultValue={resolved.map((_, i) => i.toString())}>
         {resolved.map((filter, index) => {
           switch (filter.type) {
             case 'toggle-group':
               return (
-                <Accordion
+                <Accordion.Item
                   key={index}
                   title={`${filter.label}${getParamCountLabel(params, filter.paramName)}`}
                   value={index.toString()}
                 >
                   <FilterToggleGroup paramName={filter.paramName} options={filter.options} />
-                </Accordion>
+                </Accordion.Item>
               )
             case 'range':
               return (
-                <Accordion key={index} title={filter.label} value={index.toString()}>
+                <Accordion.Item key={index} title={filter.label} value={index.toString()}>
                   <FilterRange
                     min={filter.min}
                     max={filter.max}
@@ -108,17 +108,17 @@ export function FiltersPanelInner({ className, filters }: Props) {
                     minParamName={filter.minParamName}
                     maxParamName={filter.maxParamName}
                   />
-                </Accordion>
+                </Accordion.Item>
               )
             case 'rating':
               return (
-                <Accordion key={index} title={filter.label} value={index.toString()}>
+                <Accordion.Item key={index} title={filter.label} value={index.toString()}>
                   <FilterRating paramName={filter.paramName} />
-                </Accordion>
+                </Accordion.Item>
               )
           }
         })}
-      </Accordions>
+      </Accordion.Root>
 
       <Button
         variant="secondary"
